refactor(card-browser): inline cardsListView
